+++ order/api/rest/order_handler.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"log"
