+++ order/api/rest/order_handler_test.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"bytes"
